MagentoAwareInterface concerns should be separated from steps definition
